Use size_t type for variables used to loop over arrays.

+++ src/loaders/binary.c
@@ -61,7 +61,7 @@ ansilove_binary(struct ansilove_ctx *ctx, struct ansilove_options *options)
 
 	/* process binary */
 	uint32_t character, attribute, background, foreground;
-	uint32_t loop = 0;
+	size_t loop = 0;
 	int32_t column = 0, row = 0;
 
 	while (loop < ctx->length) {
